web: move the suggestions business logic to the right module
Extract the suggestion-related logic out of the rich text module and into the suggestions module as it should have been from the beginning
